Extract grading functionality into a module
